requirements.txt
numpy>=1.24.0
torch>=2.1.0
python-chess>=1.999
requests>=2.31.0
tqdm>=4.66.0
